consolidate duplicate tests
